[Feature]: backend validations for editing and deleting adjudications, performances, awards

+++ pages/api/adjudications/delete.js
@@ -18,8 +18,37 @@ export default async (req, res) => {
     });
   }
 
+  // Check if the adjudication exists
+  const adjudication = await prisma.adjudication.findUnique({
+    where: {
+      id: id,
+    },
+  });
+
+  // If adjudication does not exist
+  if (!adjudication) {
+    return res.status(400).json({
+      error: 'Adjudication does not exist',
+    });
+  }
+
+  // Check if the performance that adjudication is associated with is finalized
+  const isFinalized = await prisma.awardPerformance.findFirst({
+    where: {
+      performance_id: adjudication.performance_id,
+      status: 'FINALIST',
+    },
+  });
+
+  // If it is finalized, we do not allow deleting adjudication
+  if (isFinalized) {
+    return res.status(400).json({
+      error: 'Adjudication cannot be deleted as the performance is finalized',
+    });
+  }
+
   // Delete adjudication
-  const deletedAdjudication = await prisma.adjudication.delete({
+  const deletedAdjudication = prisma.adjudication.delete({
     // With
     where: {
       // Specified id
@@ -27,6 +56,23 @@ export default async (req, res) => {
     },
   });
 
-  // Return deleted adjudication
-  return res.status(200).json(deletedAdjudication);
+  // Delete nominations
+  const deleteNominations = prisma.awardPerformance.deleteMany({
+    where: {
+      user_id: adjudication.user_id,
+      performance_id: adjudication.performance_id,
+    },
+  });
+
+  // Delete the adjudication and nominations made by the same user as the adjudication for the performance
+  try {
+    await prisma.$transaction([deletedAdjudication, deleteNominations]);
+  } catch {
+    return res.status(400).json({
+      error: 'Error deleting adjudication',
+    });
+  }
+
+  // Return deleted performance
+  return res.status(200).json({ message: 'Deleted adjudication successfully' });
 };
